Add integration tests for account management, including endpoints for account retrieval, email handling, and authorization policies; enhance event handler tests for account creation and engagement, and update project dependencies for testing utilities.

+++ tests/Integration/People.IntegrationTests/Grpc/PeopleServiceGetAccountTests.cs
@@ -59,7 +59,7 @@ public sealed class PeopleServiceGetAccountTests(PostgreSqlFixture postgres) : G
         var reply = await InterceptUnaryAsync(
             interceptor,
             service,
-            new AccountRequest { Id = (long)id },
+            new AccountRequest { Id = id },
             static (s, req, ctx) => s.GetAccount(req, ctx));
 
         Assert.Equal((long)id, reply.Id);
